refact: split duplication error test the most simple way
